21 feat 사용자가 보유하고있는 자산으로 주문 기능 구현

+++ Phase_4/server/src/main/java/fift/server/domain/orderdetail/OrderDetail.java
@@ -28,7 +28,7 @@ public class OrderDetail {
     private Order order;
 
     private Integer Quantity;
-    private Long total_Price;
+    private double total_Price;
 
     @Builder
     public OrderDetail(Product product,CartItem cartItem) {
